+++ examples/index.html
@@ -1,0 +1,43 @@
+<!DOCTYPE html>
+<html>
+  <head>
+    <meta charset="utf-8">
+    <title>xAPIWrapper</title>
+    <!-- <script src="../dist/xapiwrapper.min.js"></script> -->
+
+    <script src="../lib/cryptojs_v3.1.2.js"></script>
+    <script src="../src/Utils.js"></script>
+    <script src="../src/Agent.js"></script>
+    <script src="../src/Verb.js"></script>
+    <script src="../src/verbs.js"></script>
+    <script src="../src/Statement.js"></script>
+    <script src="../src/Object.js"></script>
+    <script src="../src/xAPIWrapper.js"></script>
+
+  </head>
+  <body onload='test()'>
+
+    <script type="text/javascript">
+
+    function test() {
+      ADL.XAPIWrapper.changeConfig({
+        "endpoint": "https://lrs.adlnet.gov/xapi/",
+        "user": "tom",
+        "password": "1234"
+      });
+
+
+      var stmt = new ADL.Statement(
+        new ADL.Agent(ADL.XAPIWrapper.hash("mailto:user@example.com"), 'aaron'),
+        new ADL.Verb('http://adlnet.gov/expapi/verbs/browser_testing', 'browser_testing'),
+        new ADL.Activity("act:statement_posts/browser_testing"));
+
+      stmt.timestamp = (new Date()).toISOString();
+
+      ADL.XAPIWrapper.postStatement(stmt, function(r){console.log("Success");});
+    }
+
+    </script>
+
+  </body>
+</html>
